backend_copy/requirements.txt
absl-py==0.9.0
adal==1.2.2
asgiref==3.2.3
asn1crypto==1.3.0
astor==0.7.1
astroid==2.3.3
autopep8==1.4.4
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.25
azure-core==1.3.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.48
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
ConfigArgParse==0.13.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Cython==0.29.14
Django==3.0.3
django-cors-headers==3.2.1
django-extensions==2.2.8
django-filter==2.2.0
django-nested-admin==3.2.4
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
EasyProcess==0.2.10
ecdsa==0.13
fake-useragent==0.1.11
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
grpcio==1.23.0
h5py==2.8.0
idna==2.8
inflection==0.3.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.21.0
mccabe==0.6.1
msrest==0.6.11
msrestazure==0.6.2
mxnet==1.5.0
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==20.1
Pillow==7.0.0
pipenv==2018.11.26
protobuf==3.11.3
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.9.4
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
python-jose-cryptodome==1.3.2
python-monkey-business==1.0.0
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
selenium==3.141.0
simplejson==3.17.0
six==1.14.0
soupsieve==1.9.4
SQLAlchemy==1.3.11
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.1
tensorflow-estimator==2.0.0
termcolor==1.1.0
typed-ast==1.4.1
typing==3.6.4
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
Werkzeug==0.16.1
wrapt==1.11.2
